requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file base.txt base.in
#
django==1.11.3
pygithub==1.34
pyjwt==1.5.2              # via pygithub
pytz==2017.2              # via django
